+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/HazardLowVariant.test.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import NotificationDialogVariant from './NotificationDialogVariant';
+
+const NotificationDialogInfo = () => (
+  <NotificationDialogVariant variant="hazard-low" />
+);
+
+export default NotificationDialogInfo;